requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tb-rest-client==2.7.0
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
